:truck: api: move ResultCompetenceTree from lib to src
