+++ api/lib/domain/models/index.js
@@ -2,6 +2,7 @@ import { ComplementaryCertification } from '../../../src/certification/complemen
 import { AssessmentSimulator } from '../../../src/certification/flash-certification/domain/models/AssessmentSimulator.js';
 import { FlashAssessmentAlgorithm } from '../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
 import { CertificationAttestation } from '../../../src/certification/results/domain/models/CertificationAttestation.js';
+import { ResultCompetenceTree } from '../../../src/certification/results/domain/models/ResultCompetenceTree.js';
 import { CertificationAssessmentScore } from '../../../src/certification/scoring/domain/models/CertificationAssessmentScore.js';
 import { CertificationCandidateForSupervising } from '../../../src/certification/session-management/domain/models/CertificationCandidateForSupervising.js';
 import { CertificationOfficer } from '../../../src/certification/session-management/domain/models/CertificationOfficer.js';
@@ -98,7 +99,6 @@ import { PoleEmploiSending } from './PoleEmploiSending.js';
 import { PrivateCertificate } from './PrivateCertificate.js';
 import { ReproducibilityRate } from './ReproducibilityRate.js';
 import { ResultCompetence } from './ResultCompetence.js';
-import { ResultCompetenceTree } from './ResultCompetenceTree.js';
 import { SCOCertificationCandidate } from './SCOCertificationCandidate.js';
 import { ScoringSimulation } from './ScoringSimulation.js';
 import { ScoringSimulationContext } from './ScoringSimulationContext.js';
