+++ api/tests/tooling/domain-builder/factory/build-result-competence-tree.js
@@ -1,4 +1,4 @@
-import { ResultCompetenceTree } from '../../../../lib/domain/models/ResultCompetenceTree.js';
+import { ResultCompetenceTree } from '../../../../src/certification/results/domain/models/ResultCompetenceTree.js';
 import { buildCompetenceMark } from './build-competence-mark.js';
 import { buildCompetenceTree } from './build-competence-tree.js';
