+++ api/tests/unit/infrastructure/serializers/jsonapi/private-certificate-serializer_test.js
@@ -1,6 +1,6 @@
 import { ResultCompetence } from '../../../../../lib/domain/models/ResultCompetence.js';
-import { ResultCompetenceTree } from '../../../../../lib/domain/models/ResultCompetenceTree.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/private-certificate-serializer.js';
+import { ResultCompetenceTree } from '../../../../../src/certification/results/domain/models/ResultCompetenceTree.js';
 import { AutoJuryCommentKeys } from '../../../../../src/certification/shared/domain/models/JuryComment.js';
 import { domainBuilder, expect } from '../../../../test-helper.js';
 import { getI18n } from '../../../../tooling/i18n/i18n.js';
