+++ api/tests/unit/infrastructure/serializers/jsonapi/shareable-certificate-serializer_test.js
@@ -1,6 +1,6 @@
 import { ResultCompetence } from '../../../../../lib/domain/models/ResultCompetence.js';
-import { ResultCompetenceTree } from '../../../../../lib/domain/models/ResultCompetenceTree.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/shareable-certificate-serializer.js';
+import { ResultCompetenceTree } from '../../../../../src/certification/results/domain/models/ResultCompetenceTree.js';
 import { domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | shareable-certificate-serializer', function () {
